+++ src/app/profile/profile-joint-details/profile-joint-details.component.ts
@@ -1,15 +1,15 @@
-import { Component, OnInit, ViewChild } from '@angular/core';
-import { NgForm } from '@angular/forms';
+import {Component, OnInit, ViewChild} from '@angular/core';
+import {NgForm} from '@angular/forms';
 
-import { ProfileJoint } from '../../models/profile-joint';
-import { ActivatedRoute } from '@angular/router';
-import { Addresses } from '../../models/addresses';
-import { Phones } from '../../models/phones';
-import { JointInfo } from '../../models/joint-info';
-import { Name } from '../../models/name';
-import { DateOfBirth } from '../../models/date-of-birth';
-import { Emails } from '../../models/emails';
-import { ProfileJointService } from '../joint-service';
+import {ProfileJoint} from '../../models/profile-joint';
+import {ActivatedRoute} from '@angular/router';
+import {Addresses} from '../../models/addresses';
+import {Phones} from '../../models/phones';
+import {JointInfo} from '../../models/joint-info';
+import {Name} from '../../models/name';
+import {DateOfBirth} from '../../models/date-of-birth';
+import {Emails} from '../../models/emails';
+import {ProfileJointService} from '../joint-service';
 
 @Component({
     selector: 'app-profile-joint-details',
@@ -18,7 +18,7 @@ import { ProfileJointService } from '../joint-service';
 })
 export class ProfileJointDetailsComponent implements OnInit {
     jointArray: ProfileJoint[] = [];
-    jointArray_response: ProfileJoint[] = [];
+    // jointArray_response: ProfileJoint[] = [];
     jointOneActive = false;
     jointTwoActive = false;
     // will be used to show the correct confirmation to the user
@@ -28,54 +28,52 @@ export class ProfileJointDetailsComponent implements OnInit {
     countries = this.profileJointService.countries;
     @ViewChild('f') form: NgForm;
     // get the username of the client
-    pUsername = "";
+    pUsername = '';
     // pUsername = "";
     // get the first name of the client
-    fName = "";
+    fName = '';
     // get the last name of the client
     lName = localStorage.getItem('clientLastName');
     // get the client DOB
-    DOB = "";
+    DOB = '';
     // DOB = this.jointArray['client_id']
     // get the client gender
-    pGender = "";
+    pGender = '';
     // get the client country
-    currentCountry = "";
+    currentCountry = '';
     // get the city
-    pCity = "";
+    pCity = '';
     // get the address
-    pAddress = "";
+    pAddress = '';
     // get the second address
-    pSecondAddress = "";
+    pSecondAddress = '';
     // get postal code
-    pPostalCode = "";
+    pPostalCode = '';
     // get state
-    pState = "";
+    pState = '';
     // get phone number
-    pPhoneNumber = "";
+    pPhoneNumber = '';
     // get email
-    pEmail = "";
+    pEmail = '';
 
     formData: ProfileJoint;
     private sub: any;
     joint_id: number;
-    joint_one: boolean = true
-    array_length: number
+    joint_one = true;
+    array_length: number;
 
     constructor(private profileJointService: ProfileJointService, private activatedRoute: ActivatedRoute) {
-        console.log(this.DOB)
     }
 
     ngOnInit() {
         this.sub = this.activatedRoute.params.subscribe(params => {
             this.joint_id = +params['profile_id']; // (+) converts string 'id' to a number
             if (this.joint_id === 2) {
-                this.joint_one = false
+                this.joint_one = false;
             } else {
-                this.joint_one = true
+                this.joint_one = true;
             }
             this.getDataUsingPromise();
-            console.log(this.joint_one)
             // In a real app: dispatch action to load the details here.
         });
 
@@ -97,7 +95,8 @@ export class ProfileJointDetailsComponent implements OnInit {
         //         this.pCity = this.joint_one ? jOneAddress[0].city : jSecInfo[0].city;
         //         this.pAddress = this.joint_one ? jOneAddress[0].address : jSecInfo[0].address;
         //         // // // get the second address
-        //         // // this.pSecondAddress = this.joint_one ? this.jointArray[0].name.full_name : this.jointArray[0].joint_info[0].name.full_name;
+        //         this.pSecondAddress = this.joint_one ? this.jointArray[0]
+        // .name.full_name : this.jointArray[0].joint_info[0].name.full_name;
         //         this.pPostalCode = this.joint_one ? jOneAddress[0].postal_code : jSecInfo[0].postal_code;
         //         this.pState = this.joint_one ? jOneAddress[0].state : jSecInfo[0].state;
         //         this.pPhoneNumber = jPhoneCode + '' + jPhoneNumber;
@@ -116,36 +115,35 @@ export class ProfileJointDetailsComponent implements OnInit {
 
     getDataUsingPromise() {
         return this.profileJointService.getJSON().toPromise().then(data => {
-            this.jointArray.length = 0
-            this.jointArray.push(<ProfileJoint>data['data']);
-            let jOneAddress: Addresses = this.jointArray[0].addresses;
-            let jOnePhones: Phones = this.jointArray[0].phones;
-            let jSecInfo: JointInfo = this.jointArray[0].joint_info;
-            let jName: Name = this.joint_one ? this.jointArray[0].name : jSecInfo[0].name;
-            let jDob: DateOfBirth = this.joint_one ? this.jointArray[0].date_of_birth : jSecInfo[0].date_of_birth;
-            let jOneEmail: Emails = this.jointArray[0].emails;
-            let jPhoneCode = this.joint_one ? jOnePhones[0].phone.phone_code : jSecInfo[0].phone.phone_code;
-            let jPhoneNumber = this.joint_one ? jOnePhones[0].phone.phone_number : jSecInfo[0].phone.phone_number;
-            this.fName = jName.full_name;
-            this.DOB = jDob.$date;
-            this.pGender = this.joint_one ? this.jointArray[0].gender : jSecInfo[0].gender;
-            this.currentCountry = this.joint_one ? this.jointArray[0].country : jSecInfo[0].country;
-            this.pCity = this.joint_one ? jOneAddress[0].city : jSecInfo[0].city;
-            this.pAddress = this.joint_one ? jOneAddress[0].address : jSecInfo[0].address;
-            // // // get the second address
-            // // this.pSecondAddress = this.joint_one ? this.jointArray[0].name.full_name : this.jointArray[0].joint_info[0].name.full_name;
-            this.pPostalCode = this.joint_one ? jOneAddress[0].postal_code : jSecInfo[0].postal_code;
-            this.pState = this.joint_one ? jOneAddress[0].state : jSecInfo[0].state;
-            this.pPhoneNumber = jPhoneCode + '' + jPhoneNumber;
-            this.pEmail = this.joint_one ? jOneEmail[0].email : jSecInfo[0].email;
-            this.pUsername = jOneEmail[0].email
-            console.log(this.jointArray)
-        },
+                this.jointArray.length = 0;
+                this.jointArray.push(<ProfileJoint>data['data']);
+                const jOneAddress: Addresses = this.jointArray[0].addresses;
+                const jOnePhones: Phones = this.jointArray[0].phones;
+                const jSecInfo: JointInfo = this.jointArray[0].joint_info;
+                const jName: Name = this.joint_one ? this.jointArray[0].name : jSecInfo[0].name;
+                const jDob: DateOfBirth = this.joint_one ? this.jointArray[0].date_of_birth : jSecInfo[0].date_of_birth;
+                const jOneEmail: Emails = this.jointArray[0].emails;
+                const jPhoneCode = this.joint_one ? jOnePhones[0].phone.phone_code : jSecInfo[0].phone.phone_code;
+                const jPhoneNumber = this.joint_one ? jOnePhones[0].phone.phone_number : jSecInfo[0].phone.phone_number;
+                this.fName = jName.full_name;
+                this.DOB = jDob.$date;
+                this.pGender = this.joint_one ? this.jointArray[0].gender : jSecInfo[0].gender;
+                this.currentCountry = this.joint_one ? this.jointArray[0].country : jSecInfo[0].country;
+                this.pCity = this.joint_one ? jOneAddress[0].city : jSecInfo[0].city;
+                this.pAddress = this.joint_one ? jOneAddress[0].address : jSecInfo[0].address;
+                // get the second address
+                // this.pSecondAddress = this.joint_one ?
+                // this.jointArray[0].name.full_name : this.jointArray[0].joint_info[0].name.full_name;
+                this.pPostalCode = this.joint_one ? jOneAddress[0].postal_code : jSecInfo[0].postal_code;
+                this.pState = this.joint_one ? jOneAddress[0].state : jSecInfo[0].state;
+                this.pPhoneNumber = jPhoneCode + '' + jPhoneNumber;
+                this.pEmail = this.joint_one ? jOneEmail[0].email : jSecInfo[0].email;
+                this.pUsername = jOneEmail[0].email;
+            },
             error => {
-                console.log("Error in recieving data");
+                console.log('Error in receiving data');
             }
         );
-        console.log('I will not wait until promise is resolved..');
     }
 
     onSubmit() {
@@ -161,16 +159,15 @@ export class ProfileJointDetailsComponent implements OnInit {
             this.form.value.state,
             this.form.value.phoneNumber,
             this.joint_one,
+        );
 
-        )
-
-        if (this.joint_one == true) {
+        if (this.joint_one === true) {
             this.jointOneActive = true;
         } else {
             this.jointTwoActive = true;
         }
 
-        //console.log( this.form.value.phoneNumber)
+        // console.log( this.form.value.phoneNumber)
         // this.formData = new ProfileJoint(
         // this.form.value.username,
         // this.form.value.firstName,
@@ -186,6 +183,6 @@ export class ProfileJointDetailsComponent implements OnInit {
         // this.form.value.address2
         // );
 
-        //console.log(this.formData);        
+        // console.log(this.formData);
     }
 }